tests/requirements.txt
coverage
djangocms_helper
flake8
isort
factory-boy
freezegun
mock
